solana: refactor instructions
