+++ solana/programs/token_bridge_relayer/src/constants.rs
@@ -1,0 +1,9 @@
+use anchor_lang::prelude::constant;
+
+#[constant]
+// AKA `b"bridged"`.
+pub const SEED_PREFIX_BRIDGED: &[u8; 7] = b"bridged";
+
+#[constant]
+/// AKA `b"tmp"`.
+pub const SEED_PREFIX_TMP: &[u8; 3] = b"tmp";
