--- solana/programs/token_bridge_relayer/src/context.rs
@@ -1,1073 +1,0 @@
-use anchor_lang::prelude::*;
-use anchor_spl::{
-    associated_token::AssociatedToken,
-    token::{spl_token, Mint, Token, TokenAccount},
-};
-use wormhole_anchor_sdk::{token_bridge, wormhole};
-
-use super::{
-    state::{
-        ForeignContract, OwnerConfig, RedeemerConfig, RegisteredToken, RelayerFee, SenderConfig,
-    },
-    PostedTokenBridgeRelayerMessage, TokenBridgeRelayerError,
-};
-
-// AKA `b"bridged"`.
-pub const SEED_PREFIX_BRIDGED: &[u8; 7] = b"bridged";
-/// AKA `b"tmp"`.
-pub const SEED_PREFIX_TMP: &[u8; 3] = b"tmp";
-
-#[derive(Accounts)]
-#[instruction(chain: u16)]
-pub struct RegisterForeignContract<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`SenderConfig`] account. Signer for
-    /// creating [`ForeignContract`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This program requires that the `owner` specified
-    /// in the context equals the pubkey specified in this account. Read-only.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    #[account(
-        init_if_needed,
-        payer = owner,
-        seeds = [
-            ForeignContract::SEED_PREFIX,
-            &chain.to_le_bytes()[..]
-        ],
-        bump,
-        space = 8 + ForeignContract::INIT_SPACE
-    )]
-    /// Foreign Contract account. Create this account if an emitter has not been
-    /// registered yet for this Wormhole chain ID. If there already is a
-    /// contract address saved in this account, overwrite it.
-    pub foreign_contract: Box<Account<'info, ForeignContract>>,
-
-    #[account(
-        seeds = [
-            &chain.to_be_bytes(),
-            token_bridge_foreign_endpoint.emitter_address.as_ref()
-        ],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// Token Bridge foreign endpoint. This account should really be one
-    /// endpoint per chain, but Token Bridge's PDA allows for multiple
-    /// endpoints for each chain. We store the proper endpoint for the
-    /// emitter chain.
-    pub token_bridge_foreign_endpoint: Account<'info, token_bridge::EndpointRegistration>,
-
-    /// Token Bridge program.
-    pub token_bridge_program: Program<'info, token_bridge::program::TokenBridge>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-#[instruction(chain: u16)]
-pub struct UpdateRelayerFee<'info> {
-    #[account(mut)]
-    /// Signer of the transaction. Must be the owner or assistant.
-    pub payer: Signer<'info>,
-
-    #[account(
-        seeds = [OwnerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// The owner_config is used when updating the swap rate
-    /// so that the assistant key can be used in addition to the
-    /// owner key.
-    pub owner_config: Account<'info, OwnerConfig>,
-
-    #[account(
-        init_if_needed,
-        payer = payer,
-        seeds = [
-            RelayerFee::SEED_PREFIX,
-            &chain.to_le_bytes()[..]
-        ],
-        bump,
-        space = 8 + RelayerFee::INIT_SPACE
-    )]
-    /// Relayer Fee account. This account holds the USD denominated relayer fee
-    /// for the specified `chain`. This account is used to determine the cost of
-    /// relaying a transfer to a target chain. If there already is a relayer
-    /// fee saved in this account, overwrite it.
-    pub relayer_fee: Box<Account<'info, RelayerFee>>,
-
-    #[account(
-        seeds = [
-            ForeignContract::SEED_PREFIX,
-            &chain.to_le_bytes()[..]
-        ],
-        bump
-    )]
-    /// Pass this account to verify that the specified Chain ID has a
-    /// corresponding registered foreign contract.
-    pub foreign_contract: Box<Account<'info, ForeignContract>>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-// This context is used to update both the swap_rate_precision and
-// relayer_fee_precision.
-pub struct UpdatePrecision<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`RedeemerConfig`] and [`SenderConfig`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        mut,
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [RedeemerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Redeemer Config account. This program requires that the `owner`
-    /// specified in the context equals the pubkey specified in this account.
-    /// Mutable.
-    pub redeemer_config: Box<Account<'info, RedeemerConfig>>,
-
-    #[account(
-        mut,
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This program requires that the `owner`
-    /// specified in the context equals the pubkey specified in this account.
-    /// Mutable. The `owner` check is redundant here, but we keep it as an
-    /// extra protection for future changes to the context. Mutable.
-    pub sender_config: Box<Account<'info, SenderConfig>>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct RegisterToken<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`SenderConfig`] account. Signer for
-    /// creating [`ForeignContract`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This program requires that the `owner` specified
-    /// in the context equals the pubkey specified in this account. Read-only.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    #[account(
-        init_if_needed,
-        payer = owner,
-        space = 8 + RegisteredToken::INIT_SPACE,
-        seeds = [b"mint", mint.key().as_ref()],
-        bump
-    )]
-    /// Registered Token account. This account stores information about the
-    /// token, including the swap rate and max native swap amount. Create this
-    /// account if the mint has not been registered yet. Mutable.
-    pub registered_token: Account<'info, RegisteredToken>,
-
-    /// Mint info. This is the SPL token that will be bridged over to the
-    /// foreign contract.
-    pub mint: Account<'info, Mint>,
-
-    // Token program.
-    pub token_program: Program<'info, Token>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct DeregisterToken<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`SenderConfig`] account. Signer for
-    /// creating [`ForeignContract`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This program requires that the `owner` specified
-    /// in the context equals the pubkey specified in this account. Read-only.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    /// Mint info. This is the SPL token that will be bridged over to the
-    /// foreign contract.
-    pub mint: Account<'info, Mint>,
-
-    #[account(
-        mut,
-        seeds = [b"mint", mint.key().as_ref()],
-        bump
-    )]
-    /// Registered Token account. This account stores information about the
-    /// token, including the swap rate and max native swap amount. This account
-    /// also determines if a mint is registered or not.
-    pub registered_token: Account<'info, RegisteredToken>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct ManageToken<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`SenderConfig`] account. Signer for
-    /// creating [`ForeignContract`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This program requires that the `owner` specified
-    /// in the context equals the pubkey specified in this account. Read-only.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    #[account(
-        mut,
-        seeds = [b"mint", mint.key().as_ref()],
-        bump
-    )]
-    /// Registered Token account. This account stores information about the
-    /// token, including the swap rate and max native swap amount. The program
-    /// will modify this account when the swap rate or max native swap amount
-    /// changes. Mutable.
-    pub registered_token: Account<'info, RegisteredToken>,
-
-    /// Mint info. This is the SPL token that will be bridged over to the
-    /// foreign contract.
-    pub mint: Account<'info, Mint>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct UpdateSwapRate<'info> {
-    #[account(mut)]
-    /// The signer of the transaction. Must be the owner or assistant.
-    pub payer: Signer<'info>,
-
-    #[account(
-        seeds = [OwnerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// The owner_config is used when updating the swap rate so that the
-    /// assistant key can be used in additional to the owner key.
-    pub owner_config: Account<'info, OwnerConfig>,
-
-    #[account(
-        mut,
-        seeds = [b"mint", mint.key().as_ref()],
-        bump
-    )]
-    /// Registered Token account. This account stores information about the
-    /// token, including the swap rate and max native swap amount. The program
-    /// will modify this account to update the swap rate. Mutable.
-    pub registered_token: Account<'info, RegisteredToken>,
-
-    /// Mint info. This is the SPL token that will be bridged over to the
-    /// foreign contract.
-    pub mint: Account<'info, Mint>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct PauseOutboundTransfers<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`SenderConfig`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        mut,
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This program requires that the `owner` specified
-    /// in the context equals the pubkey specified in this account. Mutable.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct ManageOwnershipTransfer<'info> {
-    #[account(mut)]
-    /// Owner of the program set in the [`OwnerConfig`] account.
-    pub owner: Signer<'info>,
-
-    #[account(
-        mut,
-        has_one = owner @ TokenBridgeRelayerError::OwnerOnly,
-        seeds = [OwnerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Owner Config account. This program requires that the `owner` specified
-    /// in the context equals the pubkey specified in this account. Mutable.
-    pub owner_config: Account<'info, OwnerConfig>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-pub struct ConfirmOwnershipTransfer<'info> {
-    #[account(mut)]
-    /// Must be the pending owner of the program set in the [`OwnerConfig`]
-    /// account.
-    pub payer: Signer<'info>,
-
-    #[account(
-        mut,
-        seeds = [OwnerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Owner Config account. This program requires that the `pending_owner`
-    /// specified in the context equals the pubkey specified in this account.
-    pub owner_config: Account<'info, OwnerConfig>,
-
-    #[account(
-        mut,
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. This instruction will update the `owner`
-    /// specified in this account to the `pending_owner` specified in the
-    /// [`OwnerConfig`] account. Mutable.
-    pub sender_config: Box<Account<'info, SenderConfig>>,
-
-    #[account(
-        mut,
-        seeds = [RedeemerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Redeemer Config account. This instruction will update the `owner`
-    /// specified in this account to the `pending_owner` specified in the
-    /// [`OwnerConfig`] account. Mutable.
-    pub redeemer_config: Box<Account<'info, RedeemerConfig>>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-}
-
-#[derive(Accounts)]
-#[instruction(
-    amount: u64,
-    to_native_token_amount: u64,
-    recipient_chain: u16,
-    recipient_address: [u8; 32],
-    batch_id: u32,
-    wrap: bool
-)]
-pub struct SendNativeTokensWithPayload<'info> {
-    /// Payer will pay Wormhole fee to transfer tokens and create temporary
-    /// token account.
-    #[account(mut)]
-    pub payer: Signer<'info>,
-
-    #[account(
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. Acts as the signer for the Token Bridge token
-    /// transfer. Read-only.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    #[account(
-        seeds = [
-            ForeignContract::SEED_PREFIX,
-            &recipient_chain.to_le_bytes()[..]
-        ],
-        bump,
-    )]
-    /// Foreign Contract account. Send tokens to the contract specified in this
-    /// account. Funnily enough, the Token Bridge program does not have any
-    /// requirements for outbound transfers for the recipient chain to be
-    /// registered. This account provides extra protection against sending
-    /// tokens to an unregistered Wormhole chain ID. Read-only.
-    pub foreign_contract: Box<Account<'info, ForeignContract>>,
-
-    #[account(mut)]
-    /// Mint info. This is the SPL token that will be bridged over to the
-    /// foreign contract. Mutable.
-    pub mint: Box<Account<'info, Mint>>,
-
-    #[account(
-        mut,
-        associated_token::mint = mint,
-        associated_token::authority = payer,
-    )]
-    /// Payer's associated token account. We may want to make this a generic
-    /// token account in the future.
-    pub from_token_account: Account<'info, TokenAccount>,
-
-    #[account(
-        seeds = [b"mint", mint.key().as_ref()],
-        bump
-    )]
-    // Registered token account for the specified mint. This account stores
-    // information about the token. Read-only.
-    pub registered_token: Box<Account<'info, RegisteredToken>>,
-
-    #[account(
-        seeds = [
-            RelayerFee::SEED_PREFIX,
-            &recipient_chain.to_le_bytes()[..]
-        ],
-        bump
-    )]
-    // Relayer fee account for the specified recipient chain. Read-only.
-    pub relayer_fee: Box<Account<'info, RelayerFee>>,
-
-    #[account(
-        init,
-        payer = payer,
-        seeds = [
-            SEED_PREFIX_TMP,
-            mint.key().as_ref(),
-        ],
-        bump,
-        token::mint = mint,
-        token::authority = config,
-    )]
-    /// Program's temporary token account. This account is created before the
-    /// instruction is invoked to temporarily take custody of the payer's
-    /// tokens. When the tokens are finally bridged out, the token account
-    /// will have zero balance and can be closed.
-    pub tmp_token_account: Box<Account<'info, TokenAccount>>,
-
-    /// Wormhole program.
-    pub wormhole_program: Program<'info, wormhole::program::Wormhole>,
-
-    /// Token Bridge program.
-    pub token_bridge_program: Program<'info, token_bridge::program::TokenBridge>,
-
-    #[account(
-        address = config.token_bridge.config @ TokenBridgeRelayerError::InvalidTokenBridgeConfig
-    )]
-    /// Token Bridge config. Read-only.
-    pub token_bridge_config: Box<Account<'info, token_bridge::Config>>,
-
-    #[account(
-        mut,
-        seeds = [mint.key().as_ref()],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// CHECK: Token Bridge custody. This is the Token Bridge program's token
-    /// account that holds this mint's balance. This account needs to be
-    /// unchecked because a token account may not have been created for this
-    /// mint yet. Mutable.
-    pub token_bridge_custody: UncheckedAccount<'info>,
-
-    #[account(
-        address = config.token_bridge.authority_signer @ TokenBridgeRelayerError::InvalidTokenBridgeAuthoritySigner
-    )]
-    /// CHECK: Token Bridge authority signer. Read-only.
-    pub token_bridge_authority_signer: UncheckedAccount<'info>,
-
-    #[account(
-        address = config.token_bridge.custody_signer @ TokenBridgeRelayerError::InvalidTokenBridgeCustodySigner
-    )]
-    /// CHECK: Token Bridge custody signer. Read-only.
-    pub token_bridge_custody_signer: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.wormhole_bridge @ TokenBridgeRelayerError::InvalidWormholeBridge,
-    )]
-    /// Wormhole bridge data. Mutable.
-    pub wormhole_bridge: Box<Account<'info, wormhole::BridgeData>>,
-
-    #[account(
-        mut,
-        seeds = [
-            SEED_PREFIX_BRIDGED,
-            &token_bridge_sequence.next_value().to_le_bytes()[..]
-        ],
-        bump,
-    )]
-    /// CHECK: Wormhole Message. Token Bridge program writes info about the
-    /// tokens transferred in this account for our program. Mutable.
-    pub wormhole_message: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.emitter @ TokenBridgeRelayerError::InvalidTokenBridgeEmitter
-    )]
-    /// CHECK: Token Bridge emitter. Mutable.
-    pub token_bridge_emitter: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.sequence @ TokenBridgeRelayerError::InvalidTokenBridgeSequence
-    )]
-    /// CHECK: Token Bridge sequence. Mutable.
-    pub token_bridge_sequence: Box<Account<'info, wormhole::SequenceTracker>>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.wormhole_fee_collector @ TokenBridgeRelayerError::InvalidWormholeFeeCollector
-    )]
-    /// Wormhole fee collector. Mutable.
-    pub wormhole_fee_collector: Box<Account<'info, wormhole::FeeCollector>>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-
-    /// Token program.
-    pub token_program: Program<'info, Token>,
-
-    /// Associated Token program.
-    pub associated_token_program: Program<'info, AssociatedToken>,
-
-    /// Clock sysvar.
-    pub clock: Sysvar<'info, Clock>,
-
-    /// Rent sysvar.
-    pub rent: Sysvar<'info, Rent>,
-}
-
-#[derive(Accounts)]
-#[instruction(vaa_hash: [u8; 32])]
-pub struct RedeemNativeTransferWithPayload<'info> {
-    #[account(mut)]
-    /// Payer will pay Wormhole fee to transfer tokens and create temporary
-    /// token account.
-    pub payer: Signer<'info>,
-
-    #[account(
-        seeds = [RedeemerConfig::SEED_PREFIX],
-        bump = config.bump
-    )]
-    /// Redeemer Config account. Acts as the Token Bridge redeemer, which signs
-    /// for the complete transfer instruction. Read-only.
-    pub config: Box<Account<'info, RedeemerConfig>>,
-
-    #[account(
-        mut,
-        associated_token::mint = mint,
-        associated_token::authority = config.fee_recipient
-    )]
-    /// Fee recipient's token account. Must be an associated token account. Mutable.
-    pub fee_recipient_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-        seeds = [
-            ForeignContract::SEED_PREFIX,
-            &vaa.emitter_chain().to_le_bytes()[..]
-        ],
-        bump,
-        constraint = foreign_contract.verify(&vaa) @ TokenBridgeRelayerError::InvalidForeignContract
-    )]
-    /// Foreign Contract account. The registered contract specified in this
-    /// account must agree with the target address for the Token Bridge's token
-    /// transfer. Read-only.
-    pub foreign_contract: Box<Account<'info, ForeignContract>>,
-
-    #[account(
-        address = vaa.data().mint()
-    )]
-    /// Mint info. This is the SPL token that will be bridged over from the
-    /// foreign contract. This must match the token address specified in the
-    /// signed Wormhole message. Read-only.
-    pub mint: Box<Account<'info, Mint>>,
-
-    #[account(
-        mut,
-        associated_token::mint = mint,
-        associated_token::authority = recipient
-    )]
-    /// Recipient associated token account. The recipient authority check
-    /// is necessary to ensure that the recipient is the intended recipient
-    /// of the bridged tokens. Mutable.
-    pub recipient_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: recipient may differ from payer if a relayer paid for this
-    /// transaction. This instruction verifies that the recipient key
-    /// passed in this context matches the intended recipient in the vaa.
-    pub recipient: UncheckedAccount<'info>,
-
-    #[account(
-        seeds = [b"mint", mint.key().as_ref()],
-        bump
-    )]
-    // Registered token account for the specified mint. This account stores
-    // information about the token. Read-only.
-    pub registered_token: Box<Account<'info, RegisteredToken>>,
-
-    #[account(
-        seeds = [b"mint", spl_token::native_mint::ID.as_ref()],
-        bump
-    )]
-    // Registered token account for the native mint. This account stores
-    // information about the token and is used for the swap rate. Read-only.
-    pub native_registered_token: Box<Account<'info, RegisteredToken>>,
-
-    #[account(
-        init,
-        payer = payer,
-        seeds = [
-            SEED_PREFIX_TMP,
-            mint.key().as_ref(),
-        ],
-        bump,
-        token::mint = mint,
-        token::authority = config
-    )]
-    /// Program's temporary token account. This account is created before the
-    /// instruction is invoked to temporarily take custody of the payer's
-    /// tokens. When the tokens are finally bridged in, the tokens will be
-    /// transferred to the destination token accounts. This account will have
-    /// zero balance and can be closed.
-    pub tmp_token_account: Box<Account<'info, TokenAccount>>,
-
-    /// Wormhole program.
-    pub wormhole_program: Program<'info, wormhole::program::Wormhole>,
-
-    /// Token Bridge program.
-    pub token_bridge_program: Program<'info, token_bridge::program::TokenBridge>,
-
-    #[account(
-        address = config.token_bridge.config @ TokenBridgeRelayerError::InvalidTokenBridgeConfig
-    )]
-    /// CHECK: Token Bridge config. Read-only.
-    pub token_bridge_config: UncheckedAccount<'info>,
-
-    #[account(
-        seeds = [
-            wormhole::SEED_PREFIX_POSTED_VAA,
-            &vaa_hash
-        ],
-        bump,
-        seeds::program = wormhole_program,
-        constraint = vaa.data().to() == *program_id || vaa.data().to() == config.key() @ TokenBridgeRelayerError::InvalidTransferToAddress,
-        constraint = vaa.data().to_chain() == wormhole::CHAIN_ID_SOLANA @ TokenBridgeRelayerError::InvalidTransferToChain,
-        constraint = vaa.data().token_chain() == wormhole::CHAIN_ID_SOLANA @ TokenBridgeRelayerError::InvalidTransferTokenChain
-    )]
-    /// Verified Wormhole message account. The Wormhole program verified
-    /// signatures and posted the account data here. Read-only.
-    pub vaa: Box<Account<'info, PostedTokenBridgeRelayerMessage>>,
-
-    #[account(mut)]
-    /// CHECK: Token Bridge claim account. It stores a boolean, whose value
-    /// is true if the bridged assets have been claimed. If the transfer has
-    /// not been redeemed, this account will not exist yet.
-    pub token_bridge_claim: UncheckedAccount<'info>,
-
-    /// CHECK: Token Bridge foreign endpoint. This account should really be one
-    /// endpoint per chain, but the PDA allows for multiple endpoints for each
-    /// chain! We store the proper endpoint for the emitter chain.
-    pub token_bridge_foreign_endpoint: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        seeds = [mint.key().as_ref()],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// CHECK: Token Bridge custody. This is the Token Bridge program's token
-    /// account that holds this mint's balance.
-    pub token_bridge_custody: Account<'info, TokenAccount>,
-
-    #[account(
-        address = config.token_bridge.custody_signer @ TokenBridgeRelayerError::InvalidTokenBridgeCustodySigner
-    )]
-    /// CHECK: Token Bridge custody signer. Read-only.
-    pub token_bridge_custody_signer: UncheckedAccount<'info>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-
-    /// Token program.
-    pub token_program: Program<'info, Token>,
-
-    /// Associated Token program.
-    pub associated_token_program: Program<'info, AssociatedToken>,
-
-    /// Rent sysvar.
-    pub rent: Sysvar<'info, Rent>,
-}
-
-#[derive(Accounts)]
-#[instruction(
-    amount: u64,
-    to_native_token_amount: u64,
-    recipient_chain: u16,
-    recipient_address: [u8; 32],
-    batch_id: u32
-)]
-pub struct SendWrappedTokensWithPayload<'info> {
-    #[account(mut)]
-    /// Payer will pay Wormhole fee to transfer tokens.
-    pub payer: Signer<'info>,
-
-    #[account(
-        seeds = [SenderConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Sender Config account. Acts as the Token Bridge sender PDA. Mutable.
-    pub config: Box<Account<'info, SenderConfig>>,
-
-    #[account(
-        seeds = [
-            ForeignContract::SEED_PREFIX,
-            &recipient_chain.to_le_bytes()[..]
-        ],
-        bump,
-    )]
-    /// Foreign Contract account. Send tokens to the contract specified in this
-    /// account. Funnily enough, the Token Bridge program does not have any
-    /// requirements for outbound transfers for the recipient chain to be
-    /// registered. This account provides extra protection against sending
-    /// tokens to an unregistered Wormhole chain ID. Read-only.
-    pub foreign_contract: Box<Account<'info, ForeignContract>>,
-
-    #[account(
-        mut,
-        seeds = [
-            token_bridge::WrappedMint::SEED_PREFIX,
-            &token_bridge_wrapped_meta.chain.to_be_bytes(),
-            &token_bridge_wrapped_meta.token_address
-        ],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// Token Bridge wrapped mint info. This is the SPL token that will be
-    /// bridged to the foreign contract. The wrapped mint PDA must agree
-    /// with the native token's metadata. Mutable.
-    pub token_bridge_wrapped_mint: Box<Account<'info, token_bridge::WrappedMint>>,
-
-    #[account(
-        mut,
-        associated_token::mint = token_bridge_wrapped_mint,
-        associated_token::authority = payer,
-    )]
-    /// Payer's associated token account. We may want to make this a generic
-    /// token account in the future.
-    pub from_token_account: Account<'info, TokenAccount>,
-
-    #[account(
-        seeds = [b"mint", token_bridge_wrapped_mint.key().as_ref()],
-        bump
-    )]
-    // Registered token account for the specified mint. This account stores
-    // information about the token. Read-only.
-    pub registered_token: Box<Account<'info, RegisteredToken>>,
-
-    #[account(
-        seeds = [
-            RelayerFee::SEED_PREFIX,
-            &recipient_chain.to_le_bytes()[..]
-        ],
-        bump
-    )]
-    // Relayer fee account for the specified recipient chain. Read-only.
-    pub relayer_fee: Box<Account<'info, RelayerFee>>,
-
-    #[account(
-        init,
-        payer = payer,
-        seeds = [
-            SEED_PREFIX_TMP,
-            token_bridge_wrapped_mint.key().as_ref(),
-        ],
-        bump,
-        token::mint = token_bridge_wrapped_mint,
-        token::authority = config,
-    )]
-    /// Program's temporary token account. This account is created before the
-    /// instruction is invoked to temporarily take custody of the payer's
-    /// tokens. When the tokens are finally bridged out, the token account
-    /// will have zero balance and can be closed.
-    pub tmp_token_account: Box<Account<'info, TokenAccount>>,
-
-    /// Wormhole program.
-    pub wormhole_program: Program<'info, wormhole::program::Wormhole>,
-
-    /// Token Bridge program.
-    pub token_bridge_program: Program<'info, token_bridge::program::TokenBridge>,
-
-    #[account(
-        seeds = [
-            token_bridge::WrappedMeta::SEED_PREFIX,
-            token_bridge_wrapped_mint.key().as_ref()
-        ],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// Token Bridge program's wrapped metadata, which stores info
-    /// about the token from its native chain:
-    ///   * Wormhole Chain ID
-    ///   * Token's native contract address
-    ///   * Token's native decimals
-    pub token_bridge_wrapped_meta: Account<'info, token_bridge::WrappedMeta>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.config @ TokenBridgeRelayerError::InvalidTokenBridgeConfig
-    )]
-    /// Token Bridge config. Mutable.
-    pub token_bridge_config: Account<'info, token_bridge::Config>,
-
-    #[account(
-        address = config.token_bridge.authority_signer @ TokenBridgeRelayerError::InvalidTokenBridgeAuthoritySigner
-    )]
-    /// CHECK: Token Bridge authority signer. Read-only.
-    pub token_bridge_authority_signer: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.wormhole_bridge @ TokenBridgeRelayerError::InvalidWormholeBridge,
-    )]
-    /// Wormhole bridge data. Mutable.
-    pub wormhole_bridge: Box<Account<'info, wormhole::BridgeData>>,
-
-    #[account(
-        mut,
-        seeds = [
-            SEED_PREFIX_BRIDGED,
-            &token_bridge_sequence.next_value().to_le_bytes()[..]
-        ],
-        bump,
-    )]
-    /// CHECK: Wormhole Message. Token Bridge program writes info about the
-    /// tokens transferred in this account.
-    pub wormhole_message: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.emitter @ TokenBridgeRelayerError::InvalidTokenBridgeEmitter
-    )]
-    /// CHECK: Token Bridge emitter. Mutable.
-    pub token_bridge_emitter: UncheckedAccount<'info>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.sequence @ TokenBridgeRelayerError::InvalidTokenBridgeSequence
-    )]
-    /// CHECK: Token Bridge sequence. Mutable.
-    pub token_bridge_sequence: Account<'info, wormhole::SequenceTracker>,
-
-    #[account(
-        mut,
-        address = config.token_bridge.wormhole_fee_collector @ TokenBridgeRelayerError::InvalidWormholeFeeCollector
-    )]
-    /// Wormhole fee collector. Mutable.
-    pub wormhole_fee_collector: Account<'info, wormhole::FeeCollector>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-
-    /// Token program.
-    pub token_program: Program<'info, Token>,
-
-    /// Associated Token program.
-    pub associated_token_program: Program<'info, AssociatedToken>,
-
-    /// Clock sysvar.
-    pub clock: Sysvar<'info, Clock>,
-
-    /// Rent sysvar.
-    pub rent: Sysvar<'info, Rent>,
-}
-
-#[derive(Accounts)]
-#[instruction(vaa_hash: [u8; 32])]
-pub struct RedeemWrappedTransferWithPayload<'info> {
-    #[account(mut)]
-    /// Payer will pay Wormhole fee to transfer tokens and create temporary
-    /// token account.
-    pub payer: Signer<'info>,
-
-    #[account(
-        seeds = [RedeemerConfig::SEED_PREFIX],
-        bump
-    )]
-    /// Redeemer Config account. Acts as the Token Bridge redeemer, which signs
-    /// for the complete transfer instruction. Read-only.
-    pub config: Box<Account<'info, RedeemerConfig>>,
-
-    #[account(
-        mut,
-        associated_token::mint = token_bridge_wrapped_mint,
-        associated_token::authority = config.fee_recipient
-    )]
-    /// Fee recipient's token account. Must be an associated token account. Mutable.
-    pub fee_recipient_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(
-        seeds = [
-            ForeignContract::SEED_PREFIX,
-            &vaa.emitter_chain().to_le_bytes()[..]
-        ],
-        bump,
-        constraint = foreign_contract.verify(&vaa) @ TokenBridgeRelayerError::InvalidForeignContract
-    )]
-    /// Foreign Contract account. The registered contract specified in this
-    /// account must agree with the target address for the Token Bridge's token
-    /// transfer. Read-only.
-    pub foreign_contract: Box<Account<'info, ForeignContract>>,
-
-    #[account(
-        mut,
-        seeds = [
-            token_bridge::WrappedMint::SEED_PREFIX,
-            &vaa.data().token_chain().to_be_bytes(),
-            vaa.data().token_address()
-        ],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// Token Bridge wrapped mint info. This is the SPL token that will be
-    /// bridged from the foreign contract. The wrapped mint PDA must agree
-    /// with the native token's metadata in the wormhole message. Mutable.
-    pub token_bridge_wrapped_mint: Box<Account<'info, token_bridge::WrappedMint>>,
-
-    #[account(
-        mut,
-        associated_token::mint = token_bridge_wrapped_mint,
-        associated_token::authority = recipient
-    )]
-    /// Recipient associated token account. The recipient authority check
-    /// is necessary to ensure that the recipient is the intended recipient
-    /// of the bridged tokens. Mutable.
-    pub recipient_token_account: Box<Account<'info, TokenAccount>>,
-
-    #[account(mut)]
-    /// CHECK: recipient may differ from payer if a relayer paid for this
-    /// transaction. This instruction verifies that the recipient key
-    /// passed in this context matches the intended recipient in the vaa.
-    pub recipient: UncheckedAccount<'info>,
-
-    #[account(
-        seeds = [b"mint", token_bridge_wrapped_mint.key().as_ref()],
-        bump
-    )]
-    // Registered token account for the specified mint. This account stores
-    // information about the token. Read-only.
-    pub registered_token: Box<Account<'info, RegisteredToken>>,
-
-    #[account(
-        seeds = [b"mint", spl_token::native_mint::ID.as_ref()],
-        bump
-    )]
-    // Registered token account for the native mint. This account stores
-    // information about the token and is used for the swap rate. Read-only.
-    pub native_registered_token: Box<Account<'info, RegisteredToken>>,
-
-    #[account(
-        init,
-        payer = payer,
-        seeds = [
-            SEED_PREFIX_TMP,
-            token_bridge_wrapped_mint.key().as_ref(),
-        ],
-        bump,
-        token::mint = token_bridge_wrapped_mint,
-        token::authority = config
-    )]
-    /// Program's temporary token account. This account is created before the
-    /// instruction is invoked to temporarily take custody of the payer's
-    /// tokens. When the tokens are finally bridged in, the tokens will be
-    /// transferred to the destination token accounts. This account will have
-    /// zero balance and can be closed.
-    pub tmp_token_account: Box<Account<'info, TokenAccount>>,
-
-    /// Wormhole program.
-    pub wormhole_program: Program<'info, wormhole::program::Wormhole>,
-
-    /// Token Bridge program.
-    pub token_bridge_program: Program<'info, token_bridge::program::TokenBridge>,
-
-    #[account(
-        seeds = [
-            token_bridge::WrappedMeta::SEED_PREFIX,
-            token_bridge_wrapped_mint.key().as_ref()
-        ],
-        bump,
-        seeds::program = token_bridge_program
-    )]
-    /// Token Bridge program's wrapped metadata, which stores info
-    /// about the token from its native chain:
-    ///   * Wormhole Chain ID
-    ///   * Token's native contract address
-    ///   * Token's native decimals
-    pub token_bridge_wrapped_meta: Box<Account<'info, token_bridge::WrappedMeta>>,
-
-    #[account(
-        address = config.token_bridge.config @ TokenBridgeRelayerError::InvalidTokenBridgeConfig
-    )]
-    /// CHECK: Token Bridge config. Read-only.
-    pub token_bridge_config: UncheckedAccount<'info>,
-
-    #[account(
-        seeds = [
-            wormhole::SEED_PREFIX_POSTED_VAA,
-            &vaa_hash
-        ],
-        bump,
-        seeds::program = wormhole_program,
-        constraint = vaa.data().to() == *program_id || vaa.data().to() == config.key() @ TokenBridgeRelayerError::InvalidTransferToAddress,
-        constraint = vaa.data().to_chain() == wormhole::CHAIN_ID_SOLANA @ TokenBridgeRelayerError::InvalidTransferToChain,
-        constraint = vaa.data().token_chain() != wormhole::CHAIN_ID_SOLANA @ TokenBridgeRelayerError::InvalidTransferTokenChain
-    )]
-    /// Verified Wormhole message account. The Wormhole program verified
-    /// signatures and posted the account data here. Read-only.
-    pub vaa: Box<Account<'info, PostedTokenBridgeRelayerMessage>>,
-
-    #[account(mut)]
-    /// CHECK: Token Bridge claim account. It stores a boolean, whose value
-    /// is true if the bridged assets have been claimed. If the transfer has
-    /// not been redeemed, this account will not exist yet.
-    pub token_bridge_claim: UncheckedAccount<'info>,
-
-    #[account(
-        address = foreign_contract.token_bridge_foreign_endpoint @ TokenBridgeRelayerError::InvalidTokenBridgeForeignEndpoint
-    )]
-    /// Token Bridge foreign endpoint. This account should really be one
-    /// endpoint per chain, but the PDA allows for multiple endpoints for each
-    /// chain! We store the proper endpoint for the emitter chain.
-    pub token_bridge_foreign_endpoint: Account<'info, token_bridge::EndpointRegistration>,
-
-    #[account(
-        address = config.token_bridge.mint_authority @ TokenBridgeRelayerError::InvalidTokenBridgeMintAuthority
-    )]
-    /// CHECK: Token Bridge custody signer. Read-only.
-    pub token_bridge_mint_authority: UncheckedAccount<'info>,
-
-    /// System program.
-    pub system_program: Program<'info, System>,
-
-    /// Token program.
-    pub token_program: Program<'info, Token>,
-
-    /// Associated Token program.
-    pub associated_token_program: Program<'info, AssociatedToken>,
-
-    /// Rent sysvar.
-    pub rent: Sysvar<'info, Rent>,
-}
