+++ solana/programs/token_bridge_relayer/src/processor/mod.rs
@@ -1,3 +1,29 @@
 mod initialize;
+mod register_foreign_contract;
+mod register_token;
+mod deregister_token;
+mod update_relayer_fee;
+mod update_precision;
+mod update_swap_rate;
+mod update_max_native_swap_amount;
+mod set_pause_for_transfers;
+mod ownership;
+mod transfer_native_tokens_with_relay;
+mod transfer_wrapped_tokens_with_relay;
+mod complete_native_transfer_with_relay;
+mod complete_wrapped_transfer_with_relay;
 
 pub use initialize::*;
+pub use register_foreign_contract::*;
+pub use register_token::*;
+pub use deregister_token::*;
+pub use update_relayer_fee::*;
+pub use update_precision::*;
+pub use update_swap_rate::*;
+pub use update_max_native_swap_amount::*;
+pub use set_pause_for_transfers::*;
+pub use ownership::*;
+pub use transfer_native_tokens_with_relay::*;
+pub use transfer_wrapped_tokens_with_relay::*;
+pub use complete_native_transfer_with_relay::*;
+pub use complete_wrapped_transfer_with_relay::*;
