+++ solana/programs/token_bridge_relayer/src/processor/update_relayer_fee.rs
@@ -1,0 +1,74 @@
+use crate::{
+    error::TokenBridgeRelayerError,
+    state::{ForeignContract, OwnerConfig, RelayerFee}
+};
+use anchor_lang::prelude::*;
+
+#[derive(Accounts)]
+#[instruction(chain: u16)]
+pub struct UpdateRelayerFee<'info> {
+    #[account(mut)]
+    /// Signer of the transaction. Must be the owner or assistant.
+    pub payer: Signer<'info>,
+
+    #[account(
+        seeds = [OwnerConfig::SEED_PREFIX],
+        bump
+    )]
+    /// The owner_config is used when updating the swap rate
+    /// so that the assistant key can be used in addition to the
+    /// owner key.
+    pub owner_config: Account<'info, OwnerConfig>,
+
+    #[account(
+        init_if_needed,
+        payer = payer,
+        seeds = [
+            RelayerFee::SEED_PREFIX,
+            &chain.to_le_bytes()[..]
+        ],
+        bump,
+        space = 8 + RelayerFee::INIT_SPACE
+    )]
+    /// Relayer Fee account. This account holds the USD denominated relayer fee
+    /// for the specified `chain`. This account is used to determine the cost of
+    /// relaying a transfer to a target chain. If there already is a relayer
+    /// fee saved in this account, overwrite it.
+    pub relayer_fee: Box<Account<'info, RelayerFee>>,
+
+    #[account(
+        seeds = [
+            ForeignContract::SEED_PREFIX,
+            &chain.to_le_bytes()[..]
+        ],
+        bump
+    )]
+    /// Pass this account to verify that the specified Chain ID has a
+    /// corresponding registered foreign contract.
+    pub foreign_contract: Box<Account<'info, ForeignContract>>,
+
+    /// System program.
+    pub system_program: Program<'info, System>,
+}
+
+pub fn update_relayer_fee(ctx: Context<UpdateRelayerFee>, chain: u16, fee: u64) -> Result<()> {
+    // Check that the signer is the owner or assistant.
+    require!(
+        ctx.accounts
+            .owner_config
+            .is_authorized(&ctx.accounts.payer.key()),
+        TokenBridgeRelayerError::OwnerOrAssistantOnly
+    );
+
+    // NOTE: We do not have to check if the chainId is valid, or if a chainId
+    // has been registered with a foreign emitter. Since the ForeignContract
+    // account is required, this means the account has been created and
+    // passed the checks required for successfully registering an emitter.
+
+    // Save the chain and fee information in the RelayerFee account.
+    let relayer_fee = &mut ctx.accounts.relayer_fee;
+    relayer_fee.chain = chain;
+    relayer_fee.fee = fee;
+
+    Ok(())
+}
