+++ solana/ts/sdk/instructions/index.ts
@@ -1,9 +1,9 @@
 export * from "./initialize";
 export * from "./registerForeignContract";
-export * from "./redeemNativeTransferWithPayload";
-export * from "./redeemWrappedTransferWithPayload";
-export * from "./sendNativeTokensWithPayload";
-export * from "./sendWrappedTokensWithPayload";
+export * from "./completeNativeTransferWithRelay";
+export * from "./completeWrappedTransferWithRelay";
+export * from "./transferNativeTokensWithRelay";
+export * from "./transferWrappedTokensWithRelay";
 export * from "./types";
 export * from "./updateRelayerFee";
 export * from "./updateSwapRate";
